Addressing windows build failures

+++ src/utils/crypto_utils/src/crypto_lib.c
@@ -339,14 +339,14 @@ done:
         goto done;
     }
 
-    rsa_N = BN_bin2bn(N, N_len, NULL);
+    rsa_N = BN_bin2bn(N, (int)N_len, NULL);
 
     if (rsa_N == NULL)
     {
         goto done;
     }
 
-    rsa_e = BN_bin2bn(e, e_len, NULL);
+    rsa_e = BN_bin2bn(e, (int)e_len, NULL);
 
     if (rsa_e == NULL)
     {
